Finishes 05-flex-modal
This one was a challenge. I have been busy since starting this course,
so the knowledge comes and go. It took me the longest time to figure out
the width:100% to nail it. I was using gaps which, while putting every-
thing in place, I knew it wasn't the idea. After finally deciding to wa-
tch all back some video courses that I remembered.